requirements.txt
streamlit==1.44.1
numpy==1.26.4
pillow==10.4.0
tensorflow-cpu==2.16.1